Move old map layer registry tests into QgsProject tests
